Make async_disconnect_handler a template class
