+++ src/server/page_server.cpp
@@ -328,78 +328,6 @@ page_server::connection_handler::remove_prior_sender_sink ()
     }
 }
 
-page_server::async_disconnect_handler::async_disconnect_handler ()
-  : m_terminate { false }
-{
-  m_thread = std::thread (&page_server::async_disconnect_handler::disconnect_loop, std::ref (*this));
-}
-
-page_server::async_disconnect_handler::~async_disconnect_handler ()
-{
-  if (m_terminate.load () && m_thread.joinable ())
-    {
-      m_thread.join ();
-    }
-  else
-    {
-      assert (m_terminate.load ());
-    }
-
-  assert (m_disconnect_queue.empty ());
-}
-
-void
-page_server::async_disconnect_handler::disconnect (connection_handler_uptr_t &&handler)
-{
-  if (!m_terminate.load ())
-    {
-      std::unique_lock<std::mutex> ulock { m_queue_mtx };
-      m_disconnect_queue.emplace (std::move (handler));
-      ulock.unlock ();
-      m_queue_cv.notify_one ();
-    }
-  else
-    {
-      // cannot ask for disconnect after termination
-      assert (false);
-    }
-}
-
-void
-page_server::async_disconnect_handler::terminate ()
-{
-  m_terminate.store (true);
-  m_queue_cv.notify_one ();
-}
-
-void
-page_server::async_disconnect_handler::disconnect_loop ()
-{
-  constexpr std::chrono::seconds one_second { 1 };
-
-  std::queue<connection_handler_uptr_t> disconnect_work_buffer;
-  while (!m_terminate.load ())
-    {
-      {
-	std::unique_lock<std::mutex> ulock { m_queue_mtx };
-	if (!m_queue_cv.wait_for (ulock, one_second,
-				  [this] { return !m_disconnect_queue.empty () || m_terminate.load (); }))
-	  {
-	    continue;
-	  }
-
-	m_disconnect_queue.swap (disconnect_work_buffer);
-      }
-
-      while (!disconnect_work_buffer.empty ())
-	{
-	  connection_handler_uptr_t &front = disconnect_work_buffer.front ();
-	  front.reset (nullptr);
-	  disconnect_work_buffer.pop ();
-	}
-    }
-}
-
 void page_server::pts_mvcc_tracker::init_oldest_active_mvccid (const std::string &pts_channel_id)
 {
   std::lock_guard<std::mutex> lockg { m_pts_oldest_active_mvccids_mtx };
